+++ test/unit/fuzz-tools.test.js
@@ -398,6 +398,7 @@ import * as milchCondyleV418 from '../../lib/milch-condyle-v418.js';
 import * as cottonMyerV419 from '../../lib/cotton-myer-v419.js';
 import * as friedmanTongueV420 from '../../lib/friedman-tongue-v420.js';
 import * as sunAcCellV421 from '../../lib/sun-ac-cell-v421.js';
+import * as sunAcFlareV422 from '../../lib/sun-ac-flare-v422.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -741,6 +742,7 @@ const MODULES = {
   'cotton-myer-v419.js': cottonMyerV419,
   'friedman-tongue-v420.js': friedmanTongueV420,
   'sun-ac-cell-v421.js': sunAcCellV421,
+  'sun-ac-flare-v422.js': sunAcFlareV422,
   'tb-testing.js': tbTesting,
 };